session06/requirements.txt
openai
python-dotenv
numpy
scipy
matplotlib